+++ .php-cs-fixer.dist.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 $header = <<<'EOF'
     This file is part of Esi\SimpleTpl.
 
-    (c) 2006 - 2024 Eric Sizemore <admin@secondversion.com>
+    (c) 2006 - 2025 Eric Sizemore <admin@secondversion.com>
 
     This file is licensed under The MIT License. For the full copyright and
     license information, please view the LICENSE.md file that was distributed